+++ community/kernel-api/src/main/java/org/neo4j/internal/kernel/api/exceptions/LabelNotFoundKernelException.java
@@ -25,7 +25,7 @@ import org.neo4j.exceptions.KernelException;
 import org.neo4j.kernel.api.exceptions.Status;
 
 public class LabelNotFoundKernelException extends KernelException {
-    public LabelNotFoundKernelException(long labelId, Exception cause) {
+    public LabelNotFoundKernelException(int labelId, Exception cause) {
         super(Status.Schema.LabelAccessFailed, cause, format("Label with id=%d not found", labelId));
     }
 }
